Worked on phone number validation, email validation and addition of payment charges
