+++ pubspec.yaml
@@ -28,12 +28,16 @@ dependencies:
   # The following adds the Cupertino Icons font to your application.
   # Use with the CupertinoIcons class for iOS style icons.
   cupertino_icons: ^1.0.2
-  fluttertoast: ^8.0.7
+  #fluttertoast: ^8.0.7
+  oktoast: ^3.0.0
   carousel_slider: ^3.0.0
   cloud_firestore: ^2.2.2
   firebase_core: ^1.3.0
   http: ^0.13.3
   webview_flutter: ^2.0.8
+  pin_code_text_field: ^1.8.0
+  intl_phone_number_input: ^0.7.0+2
+  libphonenumber: ^2.0.1
 
 dev_dependencies:
   flutter_test:
